generate brand list from theme
